+++ lib/nfg_ui/components/traits/alert.rb
@@ -5,12 +5,14 @@ module NfgUi
     module Traits
       # Access to pre-designed Alert traits
       module Alert
+        include NfgUi::Components::Utilities::Traits::TraitUtilities
+
         TRAITS = %i[tip].freeze
 
         def tip_trait
-          options[:icon] = NfgUi::DEFAULT_TIP_ICON
-          options[:theme] = NfgUi::DEFAULT_TIP_THEME
-          options[:dismissible] = false
+          maybe_update_option(:icon, value: NfgUi::DEFAULT_TIP_ICON)
+          maybe_update_option(:theme, value: NfgUi::DEFAULT_TIP_THEME)
+          maybe_update_option(:dismissible, value: false)
         end
       end
     end